Exploratory data analysis and training regression model code
The first part of the code, which includes the exploratory data analysis and training of the regression model, is complete. I still need to work on details and interpretation of the results. I uploaded it as it is now because I think you will need it to complete your parts of the HW.